Quieting a warning about small integer size on OS X.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@87142 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ macosx/plugins/DFXVideo/macsrc/PluginGLView.m
@@ -627,7 +627,8 @@ void BlitScreen16NS(unsigned char * surf,long x,long y)
 			
 			if(PSXDisplay.RGB24)
 			{
-				unsigned char * pD;unsigned int startxy;
+				unsigned char * pD;
+				size_t startxy;
 				
 				surf+=PreviousPSXDisplay.Range.x0<<2;
